Removing a bug where embedding binary files is done improperly.

+++ string2URI.m
@@ -31,7 +31,7 @@ function base64string=string2URI(filename)
             bytes=fread(fid);
             fclose(fid);
         end
-        base64string=['data:image/png;base64,' matlab.net.base64encode(bytes)];
+        base64string=['data:image/png;base64,' matlab.net.base64encode(uint8(bytes))];
     elseif(or(ext==".jpg",ext==".jpeg"))
         if(isurl)
             bytes=webread(filename,weboptions("ContentType","binary"));
@@ -40,7 +40,7 @@ function base64string=string2URI(filename)
             bytes=fread(fid);
             fclose(fid);
         end
-        base64string=['data:image/jpeg;base64,' matlab.net.base64encode(bytes)];
+        base64string=['data:image/jpeg;base64,' matlab.net.base64encode(uint8(bytes))];
     elseif(ext==".wav")
         if(isurl)
             bytes=webread(filename,weboptions("ContentType","binary"));
@@ -49,7 +49,7 @@ function base64string=string2URI(filename)
             bytes=fread(fid);
             fclose(fid);
         end
-        base64string=['data:audio/wav;base64,' matlab.net.base64encode(bytes)];
+        base64string=['data:audio/wav;base64,' matlab.net.base64encode(uint8(bytes))];
     elseif(ext==".webp")
         if(isurl)
             bytes=webread(filename,weboptions("ContentType","binary"));
@@ -58,7 +58,7 @@ function base64string=string2URI(filename)
             bytes=fread(fid);
             fclose(fid);
         end
-        base64string=['data:image/webp;base64,' matlab.net.base64encode(bytes)];
+        base64string=['data:image/webp;base64,' matlab.net.base64encode(uint8(bytes))];
     else
         base64string=filename;
     end
